pyproject.toml
[tool.poetry]
name = "chattgbot"
version = "0.1.0"
description = ""
authors = ["'eerrbis@gmail.com' <trimurs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
aiogram = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
